+++ packages/blazor-workspace/Tests/NimbleBlazor.Tests.Acceptance.Client/Program.cs
@@ -1,0 +1,13 @@
+using Microsoft.AspNetCore.Components.Web;
+using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
+
+namespace NimbleBlazor.Tests.Acceptance.Client;
+public static class Program
+{
+    public static async Task Main(string[] args)
+    {
+        var builder = WebAssemblyHostBuilder.CreateDefault(args);
+
+        await builder.Build().RunAsync();
+    }
+}